Move the renderer and its cameras out into scene.js
First module out of `web/main.js`, which is 15,680 lines and the thing this
branch is for. It goes first because it is the only section in that file that
reads no top-level name declared anywhere else: everything in it is built out of
three.js and the DOM, and every other section reads *from* it. So it sits at the
bottom of the import order and imports no part of the page back.

The 121 lines moved without a character changed, and the exports are one
statement at the foot rather than an `export` keyword on each declaration. That
is not tidiness. The proof tools mutate this tree by matching exact source text -
`syntax-check`'s anchor row holds 381 of those anchors - so a moved line that
gained two characters is a control that stops matching and says nothing about it.
Measured against that row rather than argued: 381 before, 381 after.

One binding could not simply move. `viewCamera` is declared here and assigned in
`setViewCamera` four hundred lines further down, and an importer cannot assign to
what it imports. The assignment is now `useViewCamera`, exported beside the
binding; `setViewCamera` keeps the part with an opinion - which pass draws, and
whether navigation is live - and gives up only the assignment it can no longer
make. That line was checked against the anchor set first and is not inside one.

What made this safe to attempt at all was measuring the cycles properly. Over
every reference, the file's 47 sections form one strongly-connected component of
41 sections and 14,390 lines, which reads as undecomposable and is not: a
function-call cycle between ES modules is legal, because both bodies finish
evaluating before any call runs. Rebuilt over import-time references only -
column zero, outside any function body - the graph has 13 edges and no cycles at
all. So the sections can be ordered so each module reads values only from modules
before it, and this is the first of that order.

Measured, server replaying a 284-frame 130.8MB prefix of 2026-08-02-take3.knct:

- syntax-check: web/ 5 files parsed, all 381 anchors in 14 tables match once,
  41 files, 0 failed.
- timeline-check clean: PASS, 0 failed, and its own row that the page logged no
  errors - so the module graph loads and runs in a real browser rather than
  merely parsing.
